Autoshaping Double Lever v. 1.11
This autoshaping (sign-tracking/goal-tracking) program will operate one lever as a CS+ that predicts the delivery of a liquid reinforcer. The alternate lever will operate as a CS- is not associated with reinforcement.
Supports delivery of up to 30 trials of each type (60 trials total). Sessions are run independently in each box.